+++ codex/librarian/importer/create_comics.py
@@ -281,7 +281,8 @@ class CreateComicsMixin(QueuedThread):
             try:
                 try:
                     self.status_controller.start(
-                        ImportStatusTypes.FILES_MODIFIED, name=f"({len(update_paths)})"
+                        ImportStatusTypes.FILES_MODIFIED,
+                        name=f"? / {len(update_paths)}",  # TODO fix in db
                     )
                     update_count, converted_create_paths = self._update_comics(
                         library, update_paths, all_bulk_mds
@@ -290,7 +291,8 @@ class CreateComicsMixin(QueuedThread):
                 finally:
                     self.status_controller.finish(ImportStatusTypes.FILES_MODIFIED)
                 self.status_controller.start(
-                    ImportStatusTypes.FILES_CREATED, name=f"({len(create_paths)})"
+                    ImportStatusTypes.FILES_CREATED,
+                    name=f"? / {len(create_paths)}",  # TODO fix in db
                 )
                 create_count = self._create_comics(library, create_paths, all_bulk_mds)
             finally: