+++ codex/librarian/search/remove.py
@@ -1,48 +1,51 @@
 """Search Index cleanup."""
 from time import time
+from typing import Optional
 
-from haystack.constants import DJANGO_ID
-from whoosh.query import Every, Or, Term
+from humanize import naturaldelta
 
 from codex.librarian.search.status import SearchIndexStatusTypes
+from codex.librarian.search.version import VersionMixin
 from codex.models import Comic
-from codex.worker_base import WorkerBaseMixin
+from codex.search.backend import CodexSearchBackend
 
 
-class RemoveMixin(WorkerBaseMixin):
+class RemoveMixin(VersionMixin):
     """Search Index cleanup methods."""
 
-    def _get_stale_doc_ids(self, backend):
-        """Get the stale search index pks that are no longer in the database."""
-        start = time()
-        self.status_controller.start(SearchIndexStatusTypes.SEARCH_INDEX_FIND_REMOVE, 0)
+    def _remove_stale_records(
+        self, backend: Optional[CodexSearchBackend] = None  # type: ignore
+    ):
+        """Remove records not in the database from the index."""
         try:
-            self.log.debug("Looking for stale records...")
+            start_time = time()
+            self.status_controller.start(
+                SearchIndexStatusTypes.SEARCH_INDEX_REMOVE, total=0
+            )
+            if not backend:
+                backend: CodexSearchBackend = self.engine.get_backend()  # type: ignore
+            if not backend.setup_complete:
+                backend.setup(False)
 
             database_pks = (
                 Comic.objects.all().order_by("pk").values_list("pk", flat=True)
             )
-
-            mask = Or([Term(DJANGO_ID, str(pk)) for pk in database_pks])
-
-            backend.index = backend.index.refresh()
-            with backend.index.searcher() as searcher:
-                results = searcher.search(Every(), limit=None, mask=mask)
-                stale_doc_ids = results.docs()
-            return stale_doc_ids
-        finally:
-            until = start + 1
-            self.status_controller.finish(
-                SearchIndexStatusTypes.SEARCH_INDEX_FIND_REMOVE, until=until
+            count = backend.remove_batch_pks(
+                database_pks, inverse=True, sc=self.status_controller
             )
 
-    def _remove_stale_records(self, backend):
-        """Remove records not in the database from the index."""
-        try:
-            if not backend.setup_complete:
-                backend.setup()
-            stale_doc_ids = self._get_stale_doc_ids(backend)
-            backend.remove_batch(stale_doc_ids)
+            if count:
+                elapsed_time = time() - start_time
+                elapsed = naturaldelta(elapsed_time)
+                cps = int(count / elapsed_time)
+                self.log.info(
+                    f"Removed {count} ghosts from the search index"
+                    f" in {elapsed} at {cps} per second."
+                )
+            else:
+                self.log.debug("No ghosts to remove the search index.")
         except Exception as exc:
             self.log.error("While removing stale records:")
             self.log.exception(exc)
+        finally:
+            self.status_controller.finish(SearchIndexStatusTypes.SEARCH_INDEX_REMOVE)