+++ codex/librarian/search/searchd.py
@@ -1,14 +1,15 @@
 """Haystack Search index updater."""
-from codex.librarian.search.optimize import OptimizeMixin
+from codex.librarian.search.merge import MergeMixin
 from codex.librarian.search.tasks import (
-    SearchIndexOptimizeTask,
+    SearchIndexMergeTask,
     SearchIndexRebuildIfDBChangedTask,
+    SearchIndexRemoveStaleTask,
     SearchIndexUpdateTask,
 )
 from codex.librarian.search.update import UpdateMixin
 
 
-class SearchIndexerThread(UpdateMixin, OptimizeMixin):
+class SearchIndexerThread(UpdateMixin, MergeMixin):
     """A worker to handle search index update tasks."""
 
     def process_item(self, task):
@@ -17,7 +18,9 @@ class SearchIndexerThread(UpdateMixin, OptimizeMixin):
             self._rebuild_search_index_if_db_changed()
         elif isinstance(task, SearchIndexUpdateTask):
             self._update_search_index(rebuild=task.rebuild)
-        elif isinstance(task, SearchIndexOptimizeTask):
-            self._optimize_search_index(task.force)
+        elif isinstance(task, SearchIndexMergeTask):
+            self._merge_search_index(task.optimize)
+        elif isinstance(task, SearchIndexRemoveStaleTask):
+            self._remove_stale_records()
         else:
             self.log.warning(f"Bad task sent to search index thread: {task}")